+++ packages/eslint-plugin-fiori-tools/test/rules/sap-anchor-bar-visible.test.ts
@@ -1,7 +1,16 @@
 import { RuleTester } from 'eslint';
 import anchorBarVisibleRule from '../../src/rules/sap-anchor-bar-visible.js';
 import { meta, languages } from '../../src/index.js';
-import { getManifestAsCode, setup, V4_MANIFEST, V4_MANIFEST_PATH } from '../test-helper.js';
+import {
+    getManifestAsCode,
+    setup,
+    V2_ANNOTATIONS,
+    V2_ANNOTATIONS_PATH,
+    V2_MANIFEST,
+    V2_MANIFEST_PATH,
+    V4_MANIFEST,
+    V4_MANIFEST_PATH
+} from '../test-helper.js';
 
 const ruleTester = new RuleTester({
     plugins: { ['@sap-ux/eslint-plugin-fiori-tools']: { ...meta, languages } },
@@ -18,6 +27,22 @@ ruleTester.run(TEST_NAME, anchorBarVisibleRule, {
                 name: 'anchorBarVisible not configured',
                 filename: V4_MANIFEST_PATH,
                 code: JSON.stringify(V4_MANIFEST, undefined, 2)
+            },
+            []
+        ),
+        createValidTest(
+            {
+                name: 'app type fe-v2',
+                filename: V2_MANIFEST_PATH,
+                code: JSON.stringify(V2_MANIFEST, undefined, 2)
+            },
+            []
+        ),
+        createValidTest(
+            {
+                name: 'non json source code',
+                filename: V2_ANNOTATIONS_PATH,
+                code: V2_ANNOTATIONS
             },
             []
         ),