+++ packages/abap-deploy-config-sub-generator/CHANGELOG.md
@@ -1,5 +1,107 @@
 # @sap-ux/abap-deploy-config-sub-generator
 
+## 1.0.28
+
+### Patch Changes
+
+#### Bug Fixes
+
+- Register autocomplete plugin inside getPrompts so CLI renders package autocomplete correctly [[4c34a73](https://github.com/SAP/open-ux-tools/commit/4c34a73ad972fa63eb83cf7ee2f775725e7cca71)]
+
+#### Workspace Updates
+
+- @sap-ux/abap-deploy-config-inquirer 2.0.19 → 2.0.20
+
+## 1.0.27
+
+### Patch Changes
+
+#### Release Date
+
+2026-07-01
+
+#### Workspace Updates
+
+- @sap-ux/fiori-generator-shared 1.0.15 → 1.0.16
+- @sap-ux/abap-deploy-config-inquirer 2.0.18 → 2.0.19
+- @sap-ux/deploy-config-generator-shared 1.0.15 → 1.0.16
+- @sap-ux/adp-tooling 1.0.24 → 1.0.25
+
+## 1.0.26
+
+### Patch Changes
+
+#### Release Date
+
+2026-06-30
+
+#### Workspace Updates
+
+- @sap-ux/fiori-generator-shared 1.0.14 → 1.0.15
+- @sap-ux/abap-deploy-config-inquirer 2.0.17 → 2.0.18
+- @sap-ux/deploy-config-generator-shared 1.0.14 → 1.0.15
+- @sap-ux/adp-tooling 1.0.23 → 1.0.24
+
+## 1.0.25
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/project-access 2.1.3 → 2.1.4
+- @sap-ux/abap-deploy-config-writer 1.0.10 → 1.0.11
+- @sap-ux/adp-tooling 1.0.22 → 1.0.23
+- @sap-ux/axios-extension 2.0.4 → 2.0.4
+- @sap-ux/fiori-generator-shared 1.0.13 → 1.0.14
+- @sap-ux/abap-deploy-config-inquirer 2.0.16 → 2.0.17
+- @sap-ux/deploy-config-generator-shared 1.0.13 → 1.0.14
+
+## 1.0.24
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/adp-tooling 1.0.21 → 1.0.22
+
+## 1.0.23
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/abap-deploy-config-writer 1.0.9 → 1.0.10
+- @sap-ux/axios-extension 2.0.3 → 2.0.4
+- @sap-ux/feature-toggle 1.0.2 → 1.0.3
+- @sap-ux/adp-tooling 1.0.20 → 1.0.21
+- @sap-ux/btp-utils 2.0.2 → 2.0.3
+- @sap-ux/abap-deploy-config-inquirer 2.0.15 → 2.0.16
+- @sap-ux/deploy-config-generator-shared 1.0.12 → 1.0.13
+- @sap-ux/fiori-generator-shared 1.0.12 → 1.0.13
+- @sap-ux/ui5-config 1.0.3 → 1.0.3
+
+## 1.0.22
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/project-access 2.1.2 → 2.1.3
+- @sap-ux/abap-deploy-config-writer 1.0.8 → 1.0.9
+- @sap-ux/adp-tooling 1.0.19 → 1.0.20
+- @sap-ux/axios-extension 2.0.3 → 2.0.3
+- @sap-ux/fiori-generator-shared 1.0.11 → 1.0.12
+- @sap-ux/abap-deploy-config-inquirer 2.0.14 → 2.0.15
+- @sap-ux/deploy-config-generator-shared 1.0.11 → 1.0.12
+
+## 1.0.21
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/adp-tooling 1.0.18 → 1.0.19
+
 ## 1.0.20
 
 ### Patch Changes